Add trait and interface for subject
